+++ tests/dev/script-population.R
@@ -1,6 +1,7 @@
 library(ospsuite)
 
-population <- loadPopulation("tests/data/pop_10.csv")
+popFile <- getTestDataFilePath("pop.csv")
+population <- loadPopulation(popFile)
 sim <- loadSimulation("tests/data/S1.pkml")
 
 saveSimulation(sim, "tests/data/S2.pkml")