+++ tests/testthat/test-data-importer-configuration.R
@@ -11,9 +11,9 @@ test_that("it can create a new data importer configuration", {
   expect_equal(importerConfiguration$errorUnit, NULL)
   expect_equal(importerConfiguration$measurementDimension, ospDimensions$`Concentration (molar)`)
   expect_equal(importerConfiguration$measurementUnit, "µmol/l")
-  expect_equal(importerConfiguration$measurementUnitFromColumn, FALSE)
+  expect_equal(importerConfiguration$isMeasurementUnitFromColumn, FALSE)
   expect_equal(importerConfiguration$timeUnit, ospUnits$Time$h)
-  expect_equal(importerConfiguration$timeUnitFromColumn, FALSE)
+  expect_equal(importerConfiguration$isTimeUnitFromColumn, FALSE)
   expect_equal(importerConfiguration$groupingColumns, character())
   expect_equal(importerConfiguration$sheets, character())
   expect_equal(importerConfiguration$namingPattern, "{Source}.{Sheet}")
@@ -46,12 +46,12 @@ test_that("it can set time unit from column and change column name", {
   skip_on_os("linux") # TODO enable again as soon as npoi works under linux
 
   importerConfiguration <- DataImporterConfiguration$new()
-  importerConfiguration$timeUnitFromColumn <- TRUE
-  expect_equal(importerConfiguration$timeUnitFromColumn, TRUE)
+  importerConfiguration$isTimeUnitFromColumn <- TRUE
+  expect_equal(importerConfiguration$isTimeUnitFromColumn, TRUE)
   importerConfiguration$timeUnit <- "foo"
   expect_equal(importerConfiguration$timeUnit, "foo")
 
-  importerConfiguration$timeUnitFromColumn <- FALSE
+  importerConfiguration$isTimeUnitFromColumn <- FALSE
   expect_equal(importerConfiguration$timeUnit, "h")
 
   expect_error(capture.output(print(importerConfiguration)), regexp = NA)
@@ -82,13 +82,13 @@ test_that("it can set measurement unit from column and change column name", {
   skip_on_os("linux") # TODO enable again as soon as npoi works under linux
 
   importerConfiguration <- DataImporterConfiguration$new()
-  importerConfiguration$measurementUnitFromColumn <- TRUE
-  expect_equal(importerConfiguration$measurementUnitFromColumn, TRUE)
+  importerConfiguration$isMeasurementUnitFromColumn <- TRUE
+  expect_equal(importerConfiguration$isMeasurementUnitFromColumn, TRUE)
   importerConfiguration$measurementUnit <- "foo"
   expect_equal(importerConfiguration$measurementUnit, "foo")
   expect_equal(importerConfiguration$measurementDimension, NULL)
 
-  importerConfiguration$measurementUnitFromColumn <- FALSE
+  importerConfiguration$isMeasurementUnitFromColumn <- FALSE
   expect_equal(importerConfiguration$measurementDimension, ospDimensions$`Concentration (molar)`)
   expect_equal(importerConfiguration$measurementUnit, "µmol/l")
 
@@ -126,18 +126,18 @@ test_that("it can change measurement dimension without error", {
   expect_equal(importerConfiguration$measurementUnit, "kg/l")
   expect_equal(importerConfiguration$errorUnit, NULL)
 
-  importerConfiguration$measurementUnitFromColumn <- TRUE
+  importerConfiguration$isMeasurementUnitFromColumn <- TRUE
   importerConfiguration$measurementUnit <- "measCol"
   expect_equal(importerConfiguration$measurementUnit, "measCol")
   importerConfiguration$errorUnit <- "errCol"
   expect_equal(importerConfiguration$errorUnit, NULL)
 
-  expect_equal(importerConfiguration$measurementUnitFromColumn, TRUE)
+  expect_equal(importerConfiguration$isMeasurementUnitFromColumn, TRUE)
   importerConfiguration$measurementUnit <- "foo"
   expect_equal(importerConfiguration$measurementUnit, "foo")
   expect_equal(importerConfiguration$measurementDimension, NULL)
 
-  importerConfiguration$measurementUnitFromColumn <- FALSE
+  importerConfiguration$isMeasurementUnitFromColumn <- FALSE
   expect_equal(importerConfiguration$measurementDimension, ospDimensions$`Concentration (mass)`)
   expect_equal(importerConfiguration$measurementUnit, "kg/l")
   expect_equal(importerConfiguration$errorUnit, NULL)
@@ -154,7 +154,7 @@ test_that("it can change measurement dimension with error", {
   expect_equal(importerConfiguration$measurementUnit, "kg/l")
   expect_equal(importerConfiguration$errorUnit, "kg/l")
 
-  importerConfiguration$measurementUnitFromColumn <- TRUE
+  importerConfiguration$isMeasurementUnitFromColumn <- TRUE
   importerConfiguration$errorUnit <- "errCol"
   expect_equal(importerConfiguration$errorUnit, "errCol")
 
@@ -162,7 +162,7 @@ test_that("it can change measurement dimension with error", {
   expect_equal(importerConfiguration$measurementUnit, "foo")
   expect_equal(importerConfiguration$measurementDimension, NULL)
 
-  importerConfiguration$measurementUnitFromColumn <- FALSE
+  importerConfiguration$isMeasurementUnitFromColumn <- FALSE
   expect_equal(importerConfiguration$errorUnit, "kg/l")
 
   expect_error(capture.output(print(importerConfiguration)), regexp = NA)
@@ -245,56 +245,7 @@ test_that("it can remove all sheets", {
   expect_error(capture.output(print(importerConfiguration)), regexp = NA)
 })
 
-
-context("DataImporterConfiguration from file")
-
-test_that("it can load a data importer configuration", {
-  skip_on_os("linux") # TODO enable again as soon as npoi works under linux
-
-  configurationPath <- getTestDataFilePath("dataImporterConfiguration.xml")
-  importerConfiguration <- DataImporterConfiguration$new(configurationPath)
-  expect_equal(importerConfiguration$timeColumn, "Time [h]")
-  expect_equal(importerConfiguration$errorColumn, "Error [ng/ml]")
-  expect_equal(importerConfiguration$measurementColumn, "Concentration (mass)[ng/ml]")
-  expect_equal(importerConfiguration$errorType, DataErrorType$ArithmeticStdDev)
-  expect_equal(importerConfiguration$errorUnit, "ng/ml")
-  expect_equal(importerConfiguration$measurementDimension, ospDimensions$`Concentration (mass)`)
-  expect_equal(importerConfiguration$measurementUnit, "ng/ml")
-  expect_equal(importerConfiguration$measurementUnitFromColumn, FALSE)
-  expect_equal(importerConfiguration$timeUnit, ospUnits$Time$h)
-  expect_equal(importerConfiguration$timeUnitFromColumn, FALSE)
-  # just checking for the number of grouping columns here as the sequence of column
-  # names is not obvious
-  expect_equal(length(importerConfiguration$groupingColumns), 10)
-  expect_equal(importerConfiguration$sheets, character())
-
-  expect_error(capture.output(print(importerConfiguration)), regexp = NA)
-})
-
-test_that("it can load a data importer configuration with units from columns", {
-  skip_on_os("linux") # TODO enable again as soon as npoi works under linux
-
-  configurationPath <- getTestDataFilePath("dataImporterConfiguration_UnitFromColumn.xml")
-  importerConfiguration <- DataImporterConfiguration$new(configurationPath)
-  expect_equal(importerConfiguration$timeColumn, "Time [h]")
-  expect_equal(importerConfiguration$errorColumn, "Error [ng/ml]")
-  expect_equal(importerConfiguration$measurementColumn, "Concentration (mass)[ng/ml]")
-  expect_equal(importerConfiguration$errorType, DataErrorType$ArithmeticStdDev)
-  expect_equal(importerConfiguration$errorUnit, "measurementUnit")
-  expect_equal(importerConfiguration$measurementDimension, NULL)
-  expect_equal(importerConfiguration$measurementUnit, "measurementUnit")
-  expect_equal(importerConfiguration$measurementUnitFromColumn, TRUE)
-  expect_equal(importerConfiguration$timeUnit, "TimeUnit")
-  expect_equal(importerConfiguration$timeUnitFromColumn, TRUE)
-  # just checking for the number of grouping columns here as the sequence of column
-  # names is not obvious
-  expect_equal(length(importerConfiguration$groupingColumns), 11)
-  expect_equal(importerConfiguration$sheets, c("UnitsFromColumn", "UnitsFromColumn_secondSheet"))
-
-  expect_error(capture.output(print(importerConfiguration)), regexp = NA)
-})
-
-test_that("it can change naming patter", {
+test_that("it can change naming pattern", {
   skip_on_os("linux") # TODO enable again as soon as npoi works under linux
 
   importerConfiguration <- DataImporterConfiguration$new()